+++ Mr_Steeltastic/robotcontainer.py
@@ -10,7 +10,7 @@ from subsystems.arm import Arm
 from subsystems.drivetrain import Drivetrain
 from commands.autoDock import StationCorrection
 from commands.setGrabber import SetGrabber
-from commands.moveArmCommands import MoveBackToHome, MoveBackToOrigin, MoveCubePickup, PlaceCubeMid
+from commands.armCommands import *
 from commands.miscArmCommands import ToggleArmCoast
 from guitar import Guitar
 
@@ -47,15 +47,15 @@ class RobotContainer:
         if constants.USINGGUITARCONTROLLER:
             JoystickButton(self.functionsController, Guitar.Button.kRed).whenPressed(MoveCubePickup(self.arm))
             JoystickButton(self.functionsController, Guitar.Button.kBlue).whenPressed(SetGrabber(self.arm))
-            JoystickButton(self.functionsController, Guitar.Button.kYellow).whenPressed(MoveBackToHome(self.arm))
-            JoystickButton(self.functionsController, Guitar.Button.kStar).whenPressed(MoveBackToOrigin(self.arm))
+            JoystickButton(self.functionsController, Guitar.Button.kYellow).whenPressed(MoveBackToHome(self.arm, self.arm.isInCubePlace))
+            JoystickButton(self.functionsController, Guitar.Button.kStar).whenPressed(MoveToOrigin(self.arm))
             JoystickButton(self.functionsController, Guitar.Button.kOrange).whenPressed(PlaceCubeMid(self.arm))
             JoystickButton(self.functionsController, Guitar.Button.kVol).whenPressed(ToggleArmCoast(self.arm))
         else:
             JoystickButton(self.functionsController, wpilib.XboxController.Button.kB).whenPressed(MoveCubePickup(self.arm))
             JoystickButton(self.functionsController, wpilib.XboxController.Button.kA).whenPressed(SetGrabber(self.arm))
             JoystickButton(self.functionsController, wpilib.XboxController.Button.kY).whenPressed(MoveBackToHome(self.arm))
-            JoystickButton(self.functionsController, wpilib.XboxController.Button.kX).whenPressed(MoveBackToOrigin(self.arm))
+            JoystickButton(self.functionsController, wpilib.XboxController.Button.kX).whenPressed(MoveToOrigin(self.arm))
             JoystickButton(self.functionsController,wpilib.XboxController.Button.kStart).whenPressed(PlaceCubeMid(self.arm))
 
     def getAutonomousCommand(self) -> commands2.CommandBase:
