+++ Mr_Steeltastic/simgui-window.json
@@ -5,11 +5,11 @@
   "MainWindow": {
     "GLOBAL": {
       "fps": "120",
-      "height": "1369",
+      "height": "2004",
       "maximized": "1",
       "style": "0",
       "userScale": "2",
-      "width": "2560",
+      "width": "3840",
       "xpos": "0",
       "ypos": "56"
     }
@@ -18,12 +18,12 @@
     "###FMS": {
       "Collapsed": "0",
       "Pos": "7,810",
-      "Size": "283,146"
+      "Size": "32,53"
     },
     "###Joysticks": {
       "Collapsed": "0",
       "Pos": "1787,253",
-      "Size": "796,73"
+      "Size": "32,53"
     },
     "###NetworkTables": {
       "Collapsed": "0",
@@ -38,12 +38,12 @@
     "###System Joysticks": {
       "Collapsed": "0",
       "Pos": "7,525",
-      "Size": "192,218"
+      "Size": "32,53"
     },
     "###Timing": {
       "Collapsed": "0",
       "Pos": "7,225",
-      "Size": "135,127"
+      "Size": "32,53"
     },
     "Debug##Default": {
       "Collapsed": "0",
@@ -53,7 +53,7 @@
     "Robot State": {
       "Collapsed": "0",
       "Pos": "5,20",
-      "Size": "92,99"
+      "Size": "32,53"
     }
   }
 }
